batch mm and mv
